adding enhanced debug tools for Antelope lib
