+++ src/antelope/stores/nfts.ts
@@ -17,11 +17,10 @@ import {
 } from 'src/antelope/types';
 
 import { useFeedbackStore, getAntelope, useChainStore, useEVMStore, CURRENT_CONTEXT } from 'src/antelope';
-import { createTraceFunction, isTracingAll } from 'src/antelope/stores/feedback';
 import { toRaw } from 'vue';
 import { EvmAccountModel, useAccountStore } from 'src/antelope/stores/account';
 import EVMChainSettings from 'src/antelope/chains/EVMChainSettings';
-import { errorToString } from 'src/antelope/config';
+import { createTraceFunction, errorToString } from 'src/antelope/config';
 import { truncateAddress } from 'src/antelope/stores/utils/text-utils';
 import { subscribeForTransactionReceipt } from 'src/antelope/stores/utils/trx-utils';
 
@@ -122,7 +121,6 @@ export const useNftsStore = defineStore(store_name, {
         trace: createTraceFunction(store_name),
         init: () => {
             const self = useNftsStore();
-            useFeedbackStore().setDebug(store_name, isTracingAll());
             getAntelope().events.onAccountChanged.subscribe({
                 next: async ({ label, account }) => {
                     if (label) {
