+++ src/antelope/stores/resources.ts
@@ -11,7 +11,7 @@
 
 
 import { defineStore } from 'pinia';
-import { createInitFunction, createTraceFunction } from 'src/antelope/stores/feedback';
+import { createTraceFunction } from 'src/antelope/config';
 
 export interface ResourcesState {
     __: string;
@@ -25,7 +25,6 @@ export const useResourcesStore = defineStore(store_name, {
     },
     actions: {
         trace: createTraceFunction(store_name),
-        init: createInitFunction(store_name),
     },
 });
