+++ src/antelope/stores/tokens.ts
@@ -11,9 +11,8 @@ import {
 } from 'src/antelope/types';
 import { getAntelope, useFeedbackStore, useChainStore, CURRENT_CONTEXT } from 'src/antelope';
 import { toRaw } from 'vue';
-import { errorToString } from 'src/antelope/config';
+import { createTraceFunction, errorToString } from 'src/antelope/config';
 import { filter } from 'rxjs';
-import { createTraceFunction, isTracingAll } from 'src/antelope/stores/feedback';
 import { ChainModel } from 'src/antelope/stores/chain';
 import { dateIsWithinXMinutes } from 'src/antelope/stores/utils/date-utils';
 import { getTokenPriceDataFromIndexer } from 'src/api/price';
@@ -50,7 +49,6 @@ export const useTokensStore = defineStore(store_name, {
     actions: {
         trace: createTraceFunction(store_name),
         init: () => {
-            useFeedbackStore().setDebug(store_name, isTracingAll());
             getAntelope().events.onNetworkChanged.pipe(
                 filter(e => e.label === CURRENT_CONTEXT),
             ).subscribe({
