+++ src/antelope/stores/user.ts
@@ -13,12 +13,8 @@
  */
 
 import { defineStore } from 'pinia';
-import { errorToString } from 'src/antelope/config';
+import { createTraceFunction, errorToString } from 'src/antelope/config';
 import { AccountModel } from 'src/antelope/stores/account';
-import {
-    createTraceFunction,
-    isTracingAll,
-} from 'src/antelope/stores/feedback';
 import { getAntelope, useFeedbackStore } from 'src/antelope';
 
 export type AccountList = Array<AccountId>;
@@ -84,7 +80,6 @@ export const useUserStore = defineStore(store_name, {
     actions: {
         trace: createTraceFunction(store_name),
         init: () => {
-            useFeedbackStore().setDebug(store_name, isTracingAll());
             const ant = getAntelope();
             // we want to react when the logged account changes
             ant.events.onLoggedIn.subscribe((acc: AccountModel) => useUserStore().handleAccountLoggedIn(acc));
