+++ test/jest/utils/antelope/debug.ts
@@ -1,0 +1,3 @@
+
+export const createTraceFunction = jest.fn().mockImplementation(() => jest.fn());
+export const errorToString = jest.fn().mockImplementation(e => e);
