+++ test/jest/utils/antelope/store-feedback.ts
@@ -15,14 +15,8 @@ const FeedbackStore = { ...FeedbackGetters, ...FeedbackActions };
 
 const useFeedbackStore = jest.fn().mockImplementation(() => FeedbackStore);
 
-const createTraceFunction = jest.fn().mockImplementation(() => jest.fn());
-
-const isTracingAll = jest.fn().mockImplementation(() => false);
-
 jest.mock('src/antelope/stores/feedback', () => ({
     useFeedbackStore,
-    createTraceFunction,
-    isTracingAll,
 }));
 
 export {
@@ -30,6 +24,4 @@ export {
     FeedbackGetters,
     FeedbackActions,
     useFeedbackStore,
-    createTraceFunction,
-    isTracingAll,
 };
